dev update March 1 2018

+++ backend/parallel/database/migrations/2018_02_13_050539_create_schools_table.php
@@ -24,8 +24,8 @@ class CreateSchoolsTable extends Migration
             $table->text('address');
             $table->text('city')->nullable();
             $table->string('email')->nullable();
-            $table->smallInteger('phonenumber')->nullable();
-            $table->smallInteger('mobilenumber')->nullable();
+            $table->integer('phonenumber')->nullable();
+            $table->integer('mobilenumber')->nullable();
 
             $table->text('logo')->nullable();
